M  Cargo.toml, M  src/constants.rs, M  src/engine.rs, M  src/main.rs, M  src/parse.rs, M  src/tests.rs, M  src/tt.rs, M  src/useful_macros.rs, M  src/utils.rs

+++ src/useful_macros.rs
@@ -39,14 +39,14 @@ macro_rules! get_item_unchecked_mut {
 #[macro_export]
 macro_rules! interpolate {
     ($start:expr, $end:expr, $alpha:expr) => {
-        ((1.0 - $alpha as f32) * $start as f32 + $alpha as f32 * $end as f32)
+        ((1.0 - $alpha as f64) * $start as f64 + $alpha as f64 * $end as f64)
     };
 }
 
 #[macro_export]
 macro_rules! inverse_interpolate {
     ($start:expr, $end:expr, $value:expr) => {
-        ($value as f32 - $start as f32) / ($end as f32 - $start as f32)
+        ($value as f64 - $start as f64) / ($end as f64 - $start as f64)
     };
 }
 
@@ -54,9 +54,9 @@ macro_rules! inverse_interpolate {
 macro_rules! match_interpolate {
     ($new_start:expr, $new_end:expr, $old_start:expr, $old_end:expr, $old_value:expr) => {
         interpolate!(
-            $new_start as f32,
-            $new_end as f32,
-            inverse_interpolate!($old_start as f32, $old_end as f32, $old_value as f32)
+            $new_start as f64,
+            $new_end as f64,
+            inverse_interpolate!($old_start as f64, $old_end as f64, $old_value as f64)
         )
     };
 }
